Fixing snap normal and compilation errors

+++ ihmc-footstep-planning/src/test/java/us/ihmc/footstepPlanning/steppableRegions/TerrainMapMessageToolsTest.java
@@ -26,7 +26,7 @@ public class TerrainMapMessageToolsTest
       Mat snapNormalXMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(1));
       Mat snapNormalYMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(2));
       Mat snapNormalZMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(3));
-      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(4));
+      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_32FC1, new Scalar(4));
       Mat traversabilityClassMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(5));
 
       TerrainMapTools.convertToTerrainMapData(snapNormalXMap,
@@ -45,7 +45,7 @@ public class TerrainMapMessageToolsTest
       byte[] snapNormalXMapResult = resultingData.getSnapNormalXMap();
       byte[] snapNormalYMapResult = resultingData.getSnapNormalYMap();
       byte[] snapNormalZMapResult = resultingData.getSnapNormalZMap();
-      byte[] traversabilityMapResult = resultingData.getTraversabilityScoreMap();
+      float[] traversabilityMapResult = resultingData.getTraversabilityScoreMap();
       byte[] traversabilityClassMapResult = resultingData.getTraversabilityClassMap();
 
       int totalCells = cellsPerAxis * cellsPerAxis;
@@ -76,7 +76,7 @@ public class TerrainMapMessageToolsTest
       Mat snapNormalXMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(1));
       Mat snapNormalYMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(2));
       Mat snapNormalZMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(3));
-      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(4));
+      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_32FC1, new Scalar(4));
       Mat traversabilityClassMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(5));
 
       TerrainMapTools.convertToTerrainMapData(snapNormalXMap,
@@ -120,7 +120,7 @@ public class TerrainMapMessageToolsTest
       Mat snapNormalXMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(1));
       Mat snapNormalYMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(2));
       Mat snapNormalZMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(3));
-      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(4));
+      Mat traversabilityMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_32FC1, new Scalar(4));
       Mat traversabilityClassMap = new Mat(cellsPerAxis, cellsPerAxis, opencv_core.CV_8UC1, new Scalar(5));
 
       TerrainMapTools.convertToTerrainMapData(snapNormalXMap,
